+++ test/request_list.js
@@ -496,6 +496,7 @@ describe('Apify.RequestList', () => {
 
     it('it gets correct length()', async () => {
         const sources = [
+            { url: 'https://www.example.com' },
             { url: 'https://www.ams360.com' },
             { url: 'https://www.anybus.com' },
             { url: 'https://www.anychart.com' },